requirements.txt
paho-mqtt>=2.0.0
requests>=2.32.4
SmartHomeValidation @ git+https://github.com/NadavNV/SmartHomeValidation.git@1.0.4
